azure_functions/requirements.txt
azure-functions
azure-identity
azure-storage-blob
python-dotenv
requests
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
plotly
surprise
implicit
streamlit
